First stab at JKBMS module.
It "works" in that it requests and receives data from the bms, but crashes on a 8266 trying to calculate cell voltages.
Also breaks the config page which needs 5.5 kb of ram for the html.
Running out of memory.
Deciding whether to move to an ESP32 and forego ESP8266 for this module incarnation.